✅ Move test code for contentsConverter option from test/index.ts to test/options/contentsConverter.ts
