+++ test/index.ts
@@ -5,15 +5,8 @@ import util from 'util';
 
 import netlifyPublishedDate from '../src/index';
 import { dirpath as fixtures } from './helpers/fixtures';
-import createNetlify, {
-    NetlifyDeploy,
-    requestLog2str,
-} from './helpers/netlify-mock-server';
-import { hasProp } from './helpers/utils';
-
-function getPublishedDate(deploy: NetlifyDeploy): Date {
-    return new Date(deploy.published_at || deploy.created_at);
-}
+import createNetlify, { requestLog2str } from './helpers/netlify-mock-server';
+import { getPublishedDate, hasProp } from './helpers/utils';
 
 test.serial('should add correct dates to metadata', async t => {
     const metalsmith = Metalsmith(path.join(fixtures, 'basic')).use(
@@ -432,256 +425,4 @@ test('enqueued and building deploy should be ignored', async t => {
 
     t.deepEqual(files['modified.html'].published, initialPublishedDate);
     t.deepEqual(files['modified.html'].modified, modifiedPublishedDate);
-});
-
-test('The return value of the contentsConverter() option should be used for comparison: equals value', async t => {
-    const siteID = 'equals.opt-contents-converter.index.test';
-    const currentBuildDate = new Date();
-    const metalsmith = Metalsmith(path.join(fixtures, 'basic')).use(
-        netlifyPublishedDate({
-            siteID,
-            cacheDir: null,
-            defaultDate: currentBuildDate,
-            async contentsConverter() {
-                return Buffer.from([0xf0]);
-            },
-        }),
-    );
-    const server = await createNetlify(
-        siteID,
-        [
-            {
-                key: 'initial',
-                '/initial.html': { filepath: 'initial.html' },
-                '/modified.html': Buffer.from(''),
-            },
-            {},
-            {
-                key: 'added',
-                '/added.html': { filepath: 'added.html' },
-            },
-            {
-                key: 'modified',
-                '/modified.html': { filepath: 'modified.html' },
-            },
-            {
-                key: 'last',
-            },
-        ],
-        { root: metalsmith.source() },
-    );
-
-    t.log({
-        deploys: server.deploys,
-        previews: new Map(
-            [...server.nockScope.previews.entries()].map(([url, scope]) => [
-                url,
-                scope.activeMocks(),
-            ]),
-        ),
-        requestLogs: server.requestLogs,
-    });
-
-    const initialPublishedDate = getPublishedDate(
-        server.deploys.getByKey('initial'),
-    );
-    const addedPublishedDate = getPublishedDate(
-        server.deploys.getByKey('added'),
-    );
-    const modifiedPublishedDate = getPublishedDate(
-        server.deploys.getByKey('modified'),
-    );
-    const lastPublishedDate = getPublishedDate(server.deploys.getByKey('last'));
-
-    const files = await util.promisify(metalsmith.process.bind(metalsmith))();
-    const initialPagePreviewLogs = server.requestLogs.previews.filter(
-        requestLog => requestLog.path === '/initial.html',
-    );
-    const modifiedPagePreviewLogs = server.requestLogs.previews.filter(
-        requestLog => requestLog.path === '/modified.html',
-    );
-    const addedPagePreviewLogs = server.requestLogs.previews.filter(
-        requestLog => requestLog.path === '/added.html',
-    );
-    const newPagePreviewLogs = server.requestLogs.previews.filter(
-        requestLog => requestLog.path === '/new.html',
-    );
-
-    t.log({
-        files,
-        dates: [
-            { initialPublishedDate },
-            { addedPublishedDate },
-            { modifiedPublishedDate },
-            { lastPublishedDate },
-            { currentBuildDate },
-        ],
-        requestLogs: {
-            initialPagePreviewLogs: initialPagePreviewLogs.map(requestLog2str),
-            modifiedPagePreviewLogs: modifiedPagePreviewLogs.map(
-                requestLog2str,
-            ),
-            addedPagePreviewLogs: addedPagePreviewLogs.map(requestLog2str),
-            newPagePreviewLogs: newPagePreviewLogs.map(requestLog2str),
-        },
-    });
-
-    t.deepEqual(files['initial.html'].published, initialPublishedDate);
-    t.deepEqual(files['initial.html'].modified, initialPublishedDate);
-
-    t.deepEqual(files['modified.html'].published, initialPublishedDate);
-    t.deepEqual(files['modified.html'].modified, initialPublishedDate);
-
-    t.deepEqual(files['added.html'].published, addedPublishedDate);
-    t.deepEqual(files['added.html'].modified, addedPublishedDate);
-
-    t.deepEqual(files['new.html'].published, currentBuildDate);
-    t.deepEqual(files['new.html'].modified, currentBuildDate);
-
-    t.is(
-        initialPagePreviewLogs.length,
-        server.deploys.length,
-        'If the page was deployed initial, should have requested all the previews',
-    );
-    t.is(
-        modifiedPagePreviewLogs.length,
-        server.deploys.length,
-        'If the page was deployed initial and modified midway, should have requested all the previews',
-    );
-    t.true(
-        addedPagePreviewLogs.length < server.deploys.length,
-        'If the page was deployed midway, should not have requested all previews',
-    );
-    t.is(
-        newPagePreviewLogs.length,
-        1,
-        'If the page has not been deployed yet, should have requested only the first preview',
-    );
-});
-
-test('The return value of the contentsConverter() option should be used for comparison: non-equals value', async t => {
-    let i = 0;
-    const siteID = 'non-equals.opt-contents-converter.index.test';
-    const currentBuildDate = new Date();
-    const metalsmith = Metalsmith(path.join(fixtures, 'basic')).use(
-        netlifyPublishedDate({
-            siteID,
-            cacheDir: null,
-            defaultDate: currentBuildDate,
-            async contentsConverter() {
-                return Buffer.from([0xf0, i++]);
-            },
-        }),
-    );
-    const server = await createNetlify(
-        siteID,
-        [
-            {
-                key: 'initial',
-                '/initial.html': { filepath: 'initial.html' },
-                '/modified.html': Buffer.from(''),
-            },
-            {},
-            {
-                key: 'added',
-                '/added.html': { filepath: 'added.html' },
-            },
-            {
-                key: 'modified',
-                '/modified.html': { filepath: 'modified.html' },
-            },
-            {
-                key: 'last',
-            },
-        ],
-        { root: metalsmith.source() },
-    );
-
-    t.log({
-        deploys: server.deploys,
-        previews: new Map(
-            [...server.nockScope.previews.entries()].map(([url, scope]) => [
-                url,
-                scope.activeMocks(),
-            ]),
-        ),
-        requestLogs: server.requestLogs,
-    });
-
-    const initialPublishedDate = getPublishedDate(
-        server.deploys.getByKey('initial'),
-    );
-    const addedPublishedDate = getPublishedDate(
-        server.deploys.getByKey('added'),
-    );
-    const modifiedPublishedDate = getPublishedDate(
-        server.deploys.getByKey('modified'),
-    );
-    const lastPublishedDate = getPublishedDate(server.deploys.getByKey('last'));
-
-    const files = await util.promisify(metalsmith.process.bind(metalsmith))();
-    const initialPagePreviewLogs = server.requestLogs.previews.filter(
-        requestLog => requestLog.path === '/initial.html',
-    );
-    const modifiedPagePreviewLogs = server.requestLogs.previews.filter(
-        requestLog => requestLog.path === '/modified.html',
-    );
-    const addedPagePreviewLogs = server.requestLogs.previews.filter(
-        requestLog => requestLog.path === '/added.html',
-    );
-    const newPagePreviewLogs = server.requestLogs.previews.filter(
-        requestLog => requestLog.path === '/new.html',
-    );
-
-    t.log({
-        files,
-        dates: [
-            { initialPublishedDate },
-            { addedPublishedDate },
-            { modifiedPublishedDate },
-            { lastPublishedDate },
-            { currentBuildDate },
-        ],
-        requestLogs: {
-            initialPagePreviewLogs: initialPagePreviewLogs.map(requestLog2str),
-            modifiedPagePreviewLogs: modifiedPagePreviewLogs.map(
-                requestLog2str,
-            ),
-            addedPagePreviewLogs: addedPagePreviewLogs.map(requestLog2str),
-            newPagePreviewLogs: newPagePreviewLogs.map(requestLog2str),
-        },
-    });
-
-    t.deepEqual(files['initial.html'].published, lastPublishedDate);
-    t.deepEqual(files['initial.html'].modified, currentBuildDate);
-
-    t.deepEqual(files['modified.html'].published, lastPublishedDate);
-    t.deepEqual(files['modified.html'].modified, currentBuildDate);
-
-    t.deepEqual(files['added.html'].published, lastPublishedDate);
-    t.deepEqual(files['added.html'].modified, currentBuildDate);
-
-    t.deepEqual(files['new.html'].published, currentBuildDate);
-    t.deepEqual(files['new.html'].modified, currentBuildDate);
-
-    t.is(
-        initialPagePreviewLogs.length,
-        1,
-        'Only the first preview should be requested because the page content is always determined to be different',
-    );
-    t.is(
-        modifiedPagePreviewLogs.length,
-        1,
-        'Only the first preview should be requested because the page content is always determined to be different',
-    );
-    t.is(
-        addedPagePreviewLogs.length,
-        1,
-        'Only the first preview should be requested because the page content is always determined to be different',
-    );
-    t.is(
-        newPagePreviewLogs.length,
-        1,
-        'If the page has not been deployed yet, should have requested only the first preview',
-    );
 });
